+++ src/Instance/ViewInstance/ViewInstancePane/ViewInstancePane.js
@@ -182,6 +182,9 @@ const ViewInstancePane = ({
         paneSub={paneSubTitle}
         lastMenu={lastMenu}
         actionMenu={isVersionHistoryOpen ? null : actionMenu}
+        actionMenuToggleProps={{
+          'aria-label': intl.formatMessage({ id: 'ui-inventory.actions.detailRecord' }),
+        }}
         firstMenu={(
           <PaneCloseLink
             onClick={onClose}