+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Show "Source" field in "Version history" for a user not having permissions in Central tenant. Fixes UIIN-3686.
 * Update Inventory results list's select/unselect all checkbox label. Refs UIIN-3684.
 * Integrate Button "Generate identifier" in Inventory instance records. Refs UIIN-3679.
+* Update aria-label for action menu buttons based on context. Refs UIIN-3685.
 
 ## [14.0.5](https://github.com/folio-org/ui-inventory/tree/v14.0.5) (2026-07-31)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v14.0.4...v14.0.5)